+++ contrib/opencensus-ext-azure/opencensus/ext/azure/trace_exporter/__init__.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
 import logging
 
 from opencensus.common.schedule import QueueExitEvent
@@ -146,7 +147,13 @@ class AzureExporter(BaseExporter, ProcessorMixin, TransportMixin):
             else:
                 data.type = 'INPROC'
                 data.success = True
-        # TODO: links, tracestate, tags
+        if sd.links:
+            links = []
+            for link in sd.links:
+                links.append(
+                    {"operation_Id": link.trace_id, "id": link.span_id})
+            data.properties["_MS.links"] = json.dumps(links)
+        # TODO: tracestate, tags
         for key in sd.attributes:
             # This removes redundant data from ApplicationInsights
             if key.startswith('http.'):